+++ MaterialDesignExtensions/Converters/NotNullToVisibilityConverter.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Data;
+
+namespace MaterialDesignExtensions.Converters
+{
+    public class NotNullToVisibilityConverter : IValueConverter
+    {
+        public Visibility NullValue { get; set; }
+
+        public NotNullToVisibilityConverter()
+        {
+            NullValue = Visibility.Collapsed;
+        }
+
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return value == null ? NullValue : Visibility.Visible;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return Binding.DoNothing;
+        }
+    }
+}